requirements.txt
cffi==1.15.0
gevent==21.12.0
greenlet==0.4.13
hpy==0.0.3
numpy==1.22.4
Pillow==9.1.1
python-xlib==0.31
readline==6.2.4.1
six==1.16.0
websocket==0.2.1
websocket-client==1.3.2
websockets==10.3
zope.event==4.5.0
zope.interface==5.4.0
